+++ experiment/src/org/labkey/experiment/api/ClosureQueryHelper.java
@@ -8,15 +8,11 @@ import org.labkey.api.data.BaseColumnInfo;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.CoreSchema;
-import org.labkey.api.data.DataColumn;
 import org.labkey.api.data.DbSchema;
 import org.labkey.api.data.DbScope;
-import org.labkey.api.data.DisplayColumn;
-import org.labkey.api.data.DisplayColumnFactory;
 import org.labkey.api.data.JdbcType;
 import org.labkey.api.data.MaterializedQueryHelper;
 import org.labkey.api.data.MutableColumnInfo;
-import org.labkey.api.data.RenderContext;
 import org.labkey.api.data.SQLFragment;
 import org.labkey.api.data.SqlExecutor;
 import org.labkey.api.data.TableInfo;
@@ -34,7 +30,6 @@ import org.labkey.api.query.SchemaKey;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
 import org.labkey.api.util.HeartBeat;
-import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.StringExpression;
 import org.labkey.api.view.NotFoundException;
 
@@ -169,10 +164,10 @@ public class ClosureQueryHelper
 
 
     /*
-     * This can be used to add a column directly to a exp table, or to create a column
+     * This can be used to add a column directly to an exp table, or to create a column
      * in an intermediate fake lookup table
      */
-    static MutableColumnInfo createLineageLookupColumn(final ColumnInfo fkRowId, ExpObject source, ExpObject target)
+    static MutableColumnInfo createLineageDataLookupColumn(final ColumnInfo fkRowId, ExpObject source, ExpObject target)
     {
         if (!(source instanceof ExpSampleType) && !(source instanceof ExpDataClass))
             throw new IllegalStateException();
@@ -195,58 +190,7 @@ public class ClosureQueryHelper
                 return ClosureQueryHelper.getValueSql(sourceType, sourceLsid, objectId, target);
             }
         };
-        ret.setDisplayColumnFactory(new DisplayColumnFactory()
-                {
-                    @Override
-                    public DisplayColumn createRenderer(ColumnInfo colInfo)
-                    {
-                        return new DataColumn(colInfo)
-                        {
-                            @Override
-                            public @NotNull HtmlString getFormattedHtml(RenderContext ctx)
-                            {
-                                if (getBoundColumn() != null)
-                                {
-                                    Long boundColValue = (Long) getBoundColumn().getValue(ctx);
-                                    if (boundColValue != null && boundColValue < 0)
-                                    {
-                                        return HtmlString.of("<" + (-boundColValue) + " values>");
-                                    }
-                                }
-                                return super.getFormattedHtml(ctx);
-                            }
-
-                            @Override
-                            public Object getDisplayValue(RenderContext ctx)
-                            {
-                                if (getBoundColumn() != null)
-                                {
-                                    Long boundColValue = (Long) getBoundColumn().getValue(ctx);
-                                    if (boundColValue != null && boundColValue < 0)
-                                    {
-                                        return (-boundColValue) + " values";
-                                    }
-                                }
-                                return super.getDisplayValue(ctx);
-                            }
-
-                            // Don't return a URL when we have more than one ancestor so we don't link into nothingness
-                            @Override
-                            public String renderURL(RenderContext ctx)
-                            {
-                                if (getBoundColumn() != null)
-                                {
-                                    Long boundColValue = (Long) getBoundColumn().getValue(ctx);
-                                    if (boundColValue != null && boundColValue < 0)
-                                        return null;
-                                }
-                                return super.renderURL(ctx);
-                            }
-                        };
-                    }
-
-
-                });
+        ret.setDisplayColumnFactory(colInfo -> new AncestorLookupDisplayColumn(colInfo));
         ret.setLabel(target.getName());
         UserSchema schema = Objects.requireNonNull(parentTable.getUserSchema());
         var builder = new QueryForeignKey.Builder(schema, parentTable.getContainerFilter()).table(target.getName()).key("rowid");
@@ -259,7 +203,7 @@ public class ClosureQueryHelper
                 var ret = (MutableColumnInfo) super.createLookupColumn(foreignKey, displayField);
                 if (ret != null)
                 {
-                    ret.setDisplayColumnFactory(colInfo -> new AncestorLookupDisplayColumn(foreignKey, colInfo));
+                    ret.setDisplayColumnFactory(colInfo -> new AncestorLookupDataDisplayColumn(foreignKey, colInfo));
                 }
                 return ret;
             }
@@ -679,7 +623,7 @@ public class ClosureQueryHelper
                         var target = lk.getInstance(_userSchema.getContainer(), _userSchema.getUser(), displayField);
                         if (null == target)
                             return null;
-                        return ClosureQueryHelper.createLineageLookupColumn(parent, source, target);
+                        return ClosureQueryHelper.createLineageDataLookupColumn(parent, source, target);
                     }
 
                     @Override
@@ -707,7 +651,7 @@ public class ClosureQueryHelper
             super(userSchema.getDbSchema(), "Lineage Lookup", userSchema);
             ColumnInfo wrap = new BaseColumnInfo("rowid", this, JdbcType.INTEGER);
             for (var target : type.getInstances(_userSchema.getContainer(), _userSchema.getUser()))
-                addColumn(ClosureQueryHelper.createLineageLookupColumn(wrap, source, target));
+                addColumn(ClosureQueryHelper.createLineageDataLookupColumn(wrap, source, target));
         }
 
         @Override
